refactor: auto-split test_phase3_api.py
