+++ tests/api/phase3_api/test_body_position_response_contract.py
@@ -1,0 +1,69 @@
+"""Tests for Phase 3 API: URDF/MJCF rendering, analysis tools, simulation controls.
+
+Validates Pydantic contract models and route logic for:
+- URDF model parsing and serving (#1201)
+- Analysis metrics, statistics, and export (#1203)
+- Body positioning, measurement tools (#1179)
+
+See issue #1201, #1203, #1179
+"""
+
+from __future__ import annotations
+
+import pytest
+from pydantic import ValidationError
+from src.api.models.requests import (
+    BodyPositionUpdateRequest,
+    DataExportRequest,
+    MeasurementRequest,
+)
+from src.api.models.responses import (
+    AnalysisMetricsSummary,
+    AnalysisStatisticsResponse,
+    BodyPositionResponse,
+    JointAngleDisplay,
+    MeasurementResult,
+    MeasurementToolsResponse,
+    ModelListResponse,
+    URDFJointDescriptor,
+    URDFLinkGeometry,
+    URDFModelResponse,
+)
+
+# ──────────────────────────────────────────────────────────────
+#  Contract Tests: URDF Model Responses (#1201)
+# ──────────────────────────────────────────────────────────────
+
+
+# ──────────────────────────────────────────────────────────────
+#  Contract Tests: Analysis Tools (#1203)
+# ──────────────────────────────────────────────────────────────
+
+
+# ──────────────────────────────────────────────────────────────
+#  Contract Tests: Body Positioning & Measurements (#1179)
+# ──────────────────────────────────────────────────────────────
+
+
+class TestBodyPositionResponseContract:
+    """Validate BodyPositionResponse model."""
+
+    def test_full_response(self) -> None:
+        """Complete position response."""
+        resp = BodyPositionResponse(
+            body_name="torso",
+            position=[1.0, 2.0, 3.0],
+            rotation=[0.0, 0.0, 0.0],
+            status="Position set for torso",
+        )
+        assert resp.body_name == "torso"
+
+
+# ──────────────────────────────────────────────────────────────
+#  URDF Parser Tests (#1201)
+# ──────────────────────────────────────────────────────────────
+
+
+# ──────────────────────────────────────────────────────────────
+#  Model Discovery Tests (#1201)
+# ──────────────────────────────────────────────────────────────
