+++ tests/api/phase3_api/test_model_list_response_contract.py
@@ -1,0 +1,74 @@
+"""Tests for Phase 3 API: URDF/MJCF rendering, analysis tools, simulation controls.
+
+Validates Pydantic contract models and route logic for:
+- URDF model parsing and serving (#1201)
+- Analysis metrics, statistics, and export (#1203)
+- Body positioning, measurement tools (#1179)
+
+See issue #1201, #1203, #1179
+"""
+
+from __future__ import annotations
+
+import pytest
+from pydantic import ValidationError
+from src.api.models.requests import (
+    BodyPositionUpdateRequest,
+    DataExportRequest,
+    MeasurementRequest,
+)
+from src.api.models.responses import (
+    AnalysisMetricsSummary,
+    AnalysisStatisticsResponse,
+    BodyPositionResponse,
+    JointAngleDisplay,
+    MeasurementResult,
+    MeasurementToolsResponse,
+    ModelListResponse,
+    URDFJointDescriptor,
+    URDFLinkGeometry,
+    URDFModelResponse,
+)
+
+# ──────────────────────────────────────────────────────────────
+#  Contract Tests: URDF Model Responses (#1201)
+# ──────────────────────────────────────────────────────────────
+
+
+class TestModelListResponseContract:
+    """Validate ModelListResponse contract."""
+
+    def test_list_with_models(self) -> None:
+        """List with multiple model entries."""
+        resp = ModelListResponse(
+            models=[
+                {"name": "humanoid", "format": "urdf", "path": "models/humanoid.urdf"},
+                {"name": "golfer", "format": "urdf", "path": "models/golfer.urdf"},
+            ]
+        )
+        assert len(resp.models) == 2
+
+    def test_empty_list(self) -> None:
+        """Empty model list."""
+        resp = ModelListResponse(models=[])
+        assert len(resp.models) == 0
+
+
+# ──────────────────────────────────────────────────────────────
+#  Contract Tests: Analysis Tools (#1203)
+# ──────────────────────────────────────────────────────────────
+
+
+# ──────────────────────────────────────────────────────────────
+#  Contract Tests: Body Positioning & Measurements (#1179)
+# ──────────────────────────────────────────────────────────────
+
+
+# ──────────────────────────────────────────────────────────────
+#  URDF Parser Tests (#1201)
+# ──────────────────────────────────────────────────────────────
+
+
+# ──────────────────────────────────────────────────────────────
+#  Model Discovery Tests (#1201)
+# ──────────────────────────────────────────────────────────────
